+++ Views/MainWindow.cs
@@ -114,10 +114,20 @@ namespace LibraryApp.Views
             if (user == null)
                 return;
 
-            clientService.DeleteClient(userId);
+            DialogResult result = MessageBox.Show(
+                "Вы уверены, что хотите удалить пользователя?",
+                "Подтверждение удаления",
+                MessageBoxButtons.YesNo,
+                MessageBoxIcon.Warning
+            );
 
-            // Удаляем элемент из ListView
-            usersListView.Items.Remove(selectedItem);
+            if (result == DialogResult.Yes)
+            {
+                clientService.DeleteClient(userId);
+
+                // Удаляем элемент из ListView
+                usersListView.Items.Remove(selectedItem);
+            }
         }
 
         /// <summary>
@@ -192,11 +202,22 @@ namespace LibraryApp.Views
             if (book == null)
                 return;
 
-            db.Books.Remove(book);
-            db.SaveChanges();
+            DialogResult result = MessageBox.Show(
+                "Вы уверены, что хотите удалить эту книгу?",
+                "Подтверждение удаления",
+                MessageBoxButtons.YesNo,
+                MessageBoxIcon.Warning
+            );
 
             // Удаляем элемент из ListView
-            bookList.Items.Remove(selectedItem);
+            if (result == DialogResult.Yes)
+            {
+                db.Books.Remove(book);
+                db.SaveChanges();
+
+                // Удаляем элемент из ListView
+                bookList.Items.Remove(selectedItem);
+            }
         }
         /// <summary>
         /// Очистка списка книг
@@ -379,9 +400,18 @@ namespace LibraryApp.Views
         /// </summary>
         private void DeleteDb()
         {
-            db.Database.EnsureDeleted();
-            BookListClear();
-            ClientListClear();
+            DialogResult result = MessageBox.Show(
+                "Вы уверены, что хотите удалить базу данных?",
+                "Подтверждение удаления",
+                MessageBoxButtons.YesNo,
+                MessageBoxIcon.Warning
+            );
+            if (result == DialogResult.Yes)
+            {
+                db.Database.EnsureDeleted();
+                BookListClear();
+                ClientListClear();
+            }
         }
 
         /// <summary>
